+++ compiam/rhythm/tabla_transcription/utils.py
@@ -3,7 +3,7 @@ import os
 import h5py as h5
 import numpy as np
 import torch
-
+from torch.utils import data 
 
 #Dataloader class
 class TablaDataset(data.Dataset):
